+++ core/src/main/kotlin/dev/zieger/utils/misc/IReleasable.kt
@@ -1,0 +1,9 @@
+package dev.zieger.utils.misc
+
+interface IReleasable {
+
+    /**
+     * Call this than this class is not needed anymore.
+     */
+    fun release()
+}