+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+##### 2.2.47
+
+* `OnChanged`/`Observable`: added `release()` method to stop notifying,
+when the class is not needed anymore
+
 ##### 2.2.46
 
 * JitPack foo